Fixed map indicines going out of bound
Fixed map indicines going out of bound:

When randomizing boats' locations the range was [1,10] (with rules so no boat goes out of map). This was with UI ease in mind, as people don't count from 0.
When printing the map, the boundaries were working with [0,9]. The boundaries limits change from [1,10]->[0,9] was made in Map.py methods, which should really just accept ready-made parameters.

Thus, the change was moved to Gamemaster.py that sends said parameters.

To settle UI issue, the map prints to the user locations in [1,10], but user's input (where to bomb) will be changed by the computer to [0,9] after it's sent.